move test to groupBy package
